fix(game): crandom() is double — retype f64 + audit all 27 call sites (referee catch, class F1)
Raven's crandom() macro (q_shared.h:1592) is 2.0 * (random() - 0.5):
the 2.0/0.5 literals make it double-typed, and in C it keeps the whole
surrounding expression in double until a float/int store narrows. The
port computed it in f32; 7.1% of rand draws narrow to a different f32
(exhaustive brute force over all 32768 draws). Caught by the lockstep
referee as the ent115 item-toss trDelta z divergence at live-session
frame 14282 (Drop_Item: velocity[2] += 200 + crandom() * 50) —
diagnosed from the tape alone.

Rng::crandom -> f64 (random() stays f32, widening exactly); every call
site rewritten to C's expression typing: double subexpression, narrow
at the C store point. Notables: the nextthink sites now truncate ONCE
(level.time widened into the double, matching C) instead of narrowing
early; Bowcaster's 0.2f promotes as (double)0.2f, spelled 0.2f32 as f64.

Tape-replay verification (human session, 27k frames) runs with the
task #2 harness fix.

+++ crates/mp/game/src/bg_channel/rng.rs
@@ -130,10 +130,14 @@ impl Rng {
         ((self.rand() & 0x7fff) as f32) / (0x7fff as f32)
     }
 
-    /// Raven `crandom()` macro — `2.0 * (random() - 0.5)`.
+    /// Raven `crandom()` macro — `2.0 * (random() - 0.5)`. The `2.0`/`0.5`
+    /// literals are doubles, so the macro's C type is `double`; the `f32`
+    /// `random()` widens exactly to `f64` before the subtract, and every
+    /// expression touching a call stays `double` in C until it narrows to a
+    /// float lvalue/parameter.
     /// Source: `oracle/codemp/game/q_shared.h:1592`
-    pub fn crandom(&mut self) -> f32 {
-        2.0 * (self.random() - 0.5)
+    pub fn crandom(&mut self) -> f64 {
+        2.0 * (self.random() as f64 - 0.5)
     }
 }
